Create Pauser
This is much better than putting a input('Wait") on a code line to read a console output!     This auto resumes after time and reports the line number.  Use as an import from your classes.py file Pauser as pz (example) OR have in your code and I do this to make it easy to find pz = Pauser() AND THEN anywhere in your code put pz.z20 on a line pauses fro 20 seconds.. The ,arg must begin with a z as python does not accept a number stating here ... at the end of your code put  pz.info  # this displays line number of any pausers .. this also helps you to quickly find the pause.